Rename `deep_stringify_keys` to `deep_stringify`
To make room for new methods
