+++ lib/cldr/export.rb
@@ -5,7 +5,7 @@ require "fileutils"
 require "i18n/locale/tag"
 require "core_ext/string/camelize"
 require "core_ext/string/underscore"
-require "core_ext/hash/deep_stringify_keys"
+require "core_ext/hash/deep_stringify"
 require "core_ext/hash/deep_merge"
 require "core_ext/hash/deep_prune"
 require "core_ext/hash/deep_sort"
