+++ lib/core_ext/hash/deep_stringify.rb
