+++ test/core_ext/deep_stringify_test.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require File.expand_path(File.join(File.dirname(__FILE__) + "/../test_helper"))
-require File.expand_path(File.join(File.dirname(__FILE__) + "/../../lib/core_ext/hash/deep_stringify_keys"))
+require File.expand_path(File.join(File.dirname(__FILE__) + "/../../lib/core_ext/hash/deep_stringify"))
 
 class TestDeepPrune < Test::Unit::TestCase
   test "deep_stringify_keys doesn't affect values" do
